Clean ps1 (#51)
* Clean up scripts

* scripts

+++ scripts/powershell/env.ps1
@@ -2,10 +2,12 @@ param (
     [switch]$Force
 )
 
-$backendPath = Join-Path $PSScriptRoot "..\backend"
-$frontendPath = Join-Path $PSScriptRoot "..\frontend"
+$backendPath = Join-Path $PSScriptRoot "..\..\backend"
+$workerPath = Join-Path $PSScriptRoot "..\..\backend"
+$frontendPath = Join-Path $PSScriptRoot "..\..\frontend"
 $envFilePathBackend = Join-Path $backendPath ".env"
 $envFilePathFrontend = Join-Path $frontendPath ".env"
+$envFilePathWorker = Join-Path $workerPath ".env"
 
 $envContent_frontend = @'
 ##############################################
@@ -51,6 +53,7 @@ PORT=8040
 DATABASE_URL="mysql://root:password123@localhost:3306/database"
 REDIS_URL="redis://127.0.0.1:6379"
 MONGO_URL="mongodb://localhost:27017/app"
+RABBITMQ_URL="amqp://guest:guest@localhost:5672"
 
 ##############################################
 # Security
@@ -91,6 +94,46 @@ PAYMENT_CURRENCY="CAD"
 
 '@
 
+
+$envContent_worker = @'
+##############################################
+# Configuration
+##############################################
+
+ENVIRONMENT="development"
+
+##############################################
+# Server
+##############################################
+
+FRONTEND_CLIENT="http://localhost:3040"
+
+##############################################
+# Databases
+##############################################
+
+DATABASE_URL="mysql://root:password123@localhost:3306/database"
+REDIS_URL="redis://127.0.0.1:6379"
+MONGO_URL="mongodb://localhost:27017/app"
+RABBITMQ_URL="amqp://guest:guest@localhost:5672"
+
+##############################################
+# Email (SMTP credentials)
+##############################################
+
+EMAIL_USER=""
+EMAIL_PASS=""
+
+##############################################
+# Paypal
+##############################################
+PAYPAL_CLIENT_ID="paypal-clientid"
+PAYPAL_SECRET_KEY="secret-key"
+PAYPAL_API="https://api-m.sandbox.paypal.com"
+PAYMENT_CURRENCY="CAD"
+
+'@
+
 if (-Not (Test-Path $backendPath)) {
     Write-Error "Backend folder not found at $backendPath"
     exit 1
@@ -101,6 +144,11 @@ if (-Not (Test-Path $frontendPath)) {
     exit 1
 }
 
+if (-Not (Test-Path $workerPath)) {
+    Write-Error "Worker folder not found at $workerPath"
+    exit 1
+}
+
 Set-Content -Path $envFilePathBackend -Value $envContent_backend -Encoding UTF8
 
 Write-Host ".env file has been created at $envFilePathBackend"
@@ -108,3 +156,7 @@ Write-Host ".env file has been created at $envFilePathBackend"
 Set-Content -Path $envFilePathFrontend -Value $envContent_frontend -Encoding UTF8
 
 Write-Host ".env file has been created at $envFilePathFrontend"
+
+Set-Content -Path $envFilePathWorker -Value $envContent_worker -Encoding UTF8
+
+Write-Host ".env file has been created at $envFilePathWorker"
